HHH-12666 : Move tests to org.hibernate.test.exceptionhandling package; update headers
