+++ hibernate-core/src/test/java/org/hibernate/test/exceptionhandling/TransientObjectExceptionConversionTest.java
@@ -1,19 +1,10 @@
 /*
- * Copyright 2014 JBoss Inc
+ * Hibernate, Relational Persistence for Idiomatic Java
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.exception;
+package org.hibernate.test.exceptionhandling;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
